Initial structural re-write
