+++ beacon_node/eth2-libp2p/src/lib.rs
@@ -7,7 +7,6 @@ extern crate lazy_static;
 
 pub mod behaviour;
 mod config;
-pub mod discovery;
 mod metrics;
 mod peer_manager;
 pub mod rpc;
